llm-rag-chatbot/requirements.txt
pandas
python-dotenv
openpyxl
langchain
langchain-openai
langchain-core
langchain-community
langchain-chroma
chromadb
openai
ollama
tqdm
scikit-learn
langchain-google-genai
google-auth
google-api-core
google-generativeai
streamlit
ipykernel
notebook
